// Reclutamiento/ClientApp/src/app/shared-root/models/status-process.ts
import { StateValidation } from '../enums';

export interface StatusProcess {
    id?: number;
    name: string;
    stateValidation: StateValidation;
    info?: string | null;
    userName?: string | null;
    date?: Date | null;
    description?: string | null;
    canApprove?: boolean | null;
    canDeny?: boolean | null;
    canCancel?: boolean | null;
    error?: boolean | null;
    index: number;
    currentIndex?: number;
    timelapse?: string;
    descriptionRequired?: boolean | null;
    visible?: boolean | null;
}